// src/api/models/EditarDatosOrdenRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EditarDatosOrdenRequest = {
    fechaIngreso: string;
    registradoPor: string;
    tallerBodegaDestino: string;
    inspectorId: number;
    observacionesIngreso?: string | null;
};